Add `AdobeAPIExtractor`: extraction support using Adobe PDF Extract API

+++ doc_collect.py
@@ -1,6 +1,8 @@
+from operator import le
 import os
 import re
 import sys
+import glob
 import json
 import logging
 import datetime
@@ -11,6 +13,7 @@ import retrievers
 import extractors
 
 if '-d' in sys.argv or '--debug' in sys.argv:
+    logging.basicConfig(level=logging.CRITICAL)
     retrievers.logger.setLevel(logging.DEBUG)
     extractors.logger.setLevel(logging.DEBUG)
 
@@ -29,7 +32,13 @@ avl_retrievers = {
 }
 avl_extractors = {
     'pdfminer_text': extractors.PdfminerHighLevelTextExtractor(),
-    'parsr': extractors.ParsrExtractor()
+    'parsr': extractors.ParsrExtractor(),
+    # 'parsr_extended': ,
+    'adobe_api': extractors.AdobeAPIExtractor(
+        credentials_file=os.path.join(
+            "credentials", "pdfservices-api-credentials.json"
+        )
+    )
 }
 
 def main():
@@ -53,6 +62,7 @@ def main():
     parser.add_argument('-o', '--output-dir', default='Judgments', help='output directory to store judgments')
     parser.add_argument('-J', '--omit-json', action='store_false', dest='save_json',
                         help='omit saving judgment list results as JSON')
+    parser.add_argument('--skip-existing', action='store_true', help='skip operations if results already exist')
 
     args = parser.parse_args()
 
@@ -66,42 +76,48 @@ def main():
     retriever = avl_retrievers[args.court]
 
     try:
-        print(parser.prog, ': searching judgments from ', args.court, ' ... ', sep='', end='', flush=True)
-        judgments, metadata = retriever.get_judgments(
-            args.query, page=args.page,
-            start_date=args.start_date, end_date=args.end_date
-        )
+        if args.skip_existing and (judgment_files := glob.glob("*.pdf", root_dir=output_dir)):
+            print(parser.prog, ": skipping search and downloading judgments (files exist)", sep='', flush=True)
+            judgment_files = [ os.path.join(output_dir, file) for file in judgment_files ]
+        else:
+            print(parser.prog, ': searching judgments from ', args.court, ' ... ', sep='', end='', flush=True)
+            judgments, metadata = retriever.get_judgments(
+                args.query, page=args.page,
+                start_date=args.start_date, end_date=args.end_date
+            )
 
-        if not judgments:
-            print('error', flush=True)
-            print(parser.prog, ": error: no judgments found", sep='', file=sys.stderr, flush=True)
-            sys.exit(1)
-        else: print('done', flush=True)
+            if not judgments:
+                print('error', flush=True)
+                print(parser.prog, ": error: no judgments found", sep='', file=sys.stderr, flush=True)
+                sys.exit(1)
+            else: print('done', flush=True)
 
-        print(parser.prog, ': downloading judgments to "', output_dir, '" ... ', sep='', end='', flush=True)
-        judgment_files = retriever.save_documents(judgments, output_dir=output_dir)
-        print('done', flush=True)
-
-        if args.save_json:
-            json_file = f'judgments {pathsafe(now())}.json'
-            json_file_path = os.path.join(output_dir, json_file)
-            print(parser.prog, ': saving judgment search results to ', json_file, ' ... ', sep='', end='', flush=True)
-            result = {
-                'meta': {
-                    'directory': os.path.abspath(output_dir),
-                    'request': {
-                        'query': args.query or "",
-                        'court': args.court,
-                        'page': args.page or 0,
-                    },
-                },
-                'data': judgments
-            }
-            if metadata is not None:
-                result['meta']['response'] = metadata
-            with open(json_file_path, 'w+', encoding='utf-8') as file:
-                json.dump(result, file, indent=4, ensure_ascii=False)
+            print(parser.prog, ': downloading judgments to "', output_dir, '" ... ', sep='', end='', flush=True)
+            judgment_files = retriever.save_documents(judgments, output_dir=output_dir)
             print('done', flush=True)
+
+            if args.save_json:
+                json_file = f'judgments {pathsafe(now())}.json'
+                json_file_path = os.path.join(output_dir, json_file)
+                print(parser.prog, ': saving judgment search results to ', json_file, ' ... ', sep='', end='', flush=True)
+                result = {
+                    'meta': {
+                        'directory': os.path.abspath(output_dir),
+                        'request': {
+                            'query'     : args.query or "",
+                            'court'     : args.court,
+                            'page'      : args.page or 0,
+                            'start_date': args.start_date,
+                            'end_date'  : args.end_date
+                        },
+                    },
+                    'data': judgments
+                }
+                if metadata is not None:
+                    result['meta']['response'] = metadata
+                with open(json_file_path, 'w+', encoding='utf-8') as file:
+                    json.dump(result, file, indent=4, ensure_ascii=False)
+                print('done', flush=True)
 
         for extrctr in args.extractor:
             extractor = avl_extractors[extrctr]
@@ -112,7 +128,13 @@ def main():
                 extrctr, ' ... ', sep='', end='', flush=True
             )
             for pdf_file in judgment_files:
-                extractor.extract_to_file(pdf_file, output_dir=extract_output_dir)
+                try:
+                    extractor.extract_to_file(
+                        pdf_file, output_dir=extract_output_dir,
+                        skip_existing=args.skip_existing
+                    )
+                except Exception as exc:
+                    print('\n  error:', exc, ' ... ', sep='', end='')
             print('done', flush=True)
     except Exception as exc:
         print('error', flush=True)
